stripping out request handler and multi-core initial test and putting collections in the resource index

+++ search/src/main/java/org/tdar/search/service/index/BatchIndexer.java
@@ -3,8 +3,10 @@ package org.tdar.search.service.index;
 import java.io.IOException;
 import java.io.Serializable;
 import java.util.HashMap;
+import java.util.HashSet;
 import java.util.List;
 import java.util.Map;
+import java.util.Set;
 
 import org.apache.commons.lang3.StringUtils;
 import org.apache.solr.client.solrj.SolrServerException;
@@ -87,8 +89,12 @@ public class BatchIndexer implements Serializable {
             }
             Counter counter = new Counter();
             counter.setTotal(total);
+            Set<String> seenCores = new HashSet<>();
             for (LookupSource src : sources) {
-                searchIndexService.purgeCore(src.getCoreName());
+                if (!seenCores.contains(src.getCoreName())) {
+                    searchIndexService.purgeCore(src.getCoreName());
+                    seenCores.add(src.getCoreName());
+                }
 	            for (Class<? extends Indexable> toIndex : src.getClasses()) {
 	                updateAllStatuses(updateReceiver, activity, "initializing... ["+toIndex.getSimpleName()+": "+total+"]", counter.getPercent());
 	                ScrollableResults scrollableResults = null;
